requirements.txt
Flask_sqlalchemy-3.1.1 
greenlet-3.1.1 
sqlalchemy-2.0.36 
typing-extensions-4.12.2
Flask-3.1.0 
Jinja2-3.1.5 
MarkupSafe-3.0.2 
Werkzeug-3.1.3 
blinker-1.9.0 
click-8.1.8 
colorama-0.4.6 
itsdangerous-2.2.0
Flask-Migrate-4.0.7 
Mako-1.3.8 
alembic-1.14.0